requirements.txt
Pillow>=10.0.0
Django>=4.2.0
requests>=2.31.0


